Updated openstack/openstack Project: openstack-infra/project-config  abbad5ca8a49111a317b6ca0594c5495b92cfc76
Turn debug logging on for beaker jobs

Inexplicably, the current version of beaker defaults to not verbosely
logging its actions. This is a problem when run in a CI environment
because we always want as much information as possible. Setting the
environment variable BEAKER_debug to 'yes' returns the sane behavior.
This variable is documented here[1].

[1] https://github.com/puppetlabs/beaker/wiki/How-to-Write-a-Beaker-Test-for-a-Module#supported-env-variables

Change-Id: Ifd8450803bc8ab8780e81fa74feb0f455ce6be13