Updated openstack/openstack Project: openstack/cinder  69f28d96e4aef96e8b72ca421de46b260a3486e4
Add 'source-id' and 'source-name' support in VNX driver

Currently, VNX driver only supports 'id' as the id-type in manage/unmanage volume.
With release of new version python-cinderclient, 'source-id' and 'source-name' are
preferred id-type when using cinder manage.

This fix is to add 'source-id' and 'source-name' support
and obsolete the self-defined 'id' support.

Change-Id: Ia0619123ff56e44da98c4f12c554474c8fe6093c
closes-bug: 1450280