Updated openstack/openstack Project: openstack/keystone  dc15388ffac0425db77f46c9d13823deb25cd81f
Merge tag '2015.1.0'

This is a null-merge of the 2015.1.0 release tag back into the master
branch so that the 2015.1.0 tag will appear in the git commit history of
the master branch. It contains no actual changes to the master branch,
regardless of how our code review system's UI represents it. Please
ask in #openstack-infra if you have any questions, and otherwise try
to merge this as quickly as possible to avoid later conflicts on the
master branch.

Change-Id: I7467bea4c34a852b4efd9ebb8178aa8c77e9a7f7