Updated openstack/openstack Project: openstack/keystone  b3c3022354a631bf7ea42e5bcbf980c5f14925e1
Merge tag '2014.2'

This is a null-merge of the 2014.2 release tag back into the
master branch. It contains no actual changes to the master branch,
regardless of how our code review system's UI represents it. Please
ask in #openstack-infra if you have any questions, and otherwise try
to merge this as quickly as possible to avoid later conflicts on the
master branch.

Change-Id: I46c0f42fde056803278f44797375812d605c2af0