Updated openstack/openstack Project: openstack/trove-integration  9034464143c05e9b97399c23544df147b81a21b5
Enhance 'stop' and 'start' calls

If you run 'redstack start' no check is made to see if the processes
are already running.  This can potentially cause multiple sets of
the Trove services to be running at the same time.

By the same token, calling 'stop' when several sets are running will
only remove one set, potentially leaving multiple sets still executing.

'start' now checks to make sure that no Trove process are running; if
so, a warning is emitted.  'stop' tries to remove up to 5 instances
of Trove processes.  If more than 5 are found a warning is emitted.
(5 was chosen arbitrarily so that failing to stop the processes
wouldn't result in an endless loop.)

A 'restart' command was also added, which just calls 'stop' then
'start.'

Change-Id: I4df6490c437deecfe3c2723d5e7965b2114adc9f