Updated openstack/openstack Project: openstack/neutron  718e7714f52bd953e70770bd1d5e20b4ce4412a7
Merge tag '2015.1.0'

This is a null-merge of the 2015.1.0 release tag back into the master
branch so that the 2015.1.0 tag will appear in the git commit history of
the master branch. It contains no actual changes to the master branch,
regardless of how our code review system's UI represents it. Please
ask in #openstack-infra if you have any questions, and otherwise try
to merge this as quickly as possible to avoid later conflicts on the
master branch.

Change-Id: I69e520148633c8f1ec1be019ec83f06f11b84060