Updated openstack/openstack Project: openstack/governance  0886e901287a6c664cdabc54f2fe68d2fd36faef
Add a reference for User Committee repos

Currently the User Committee doesn't have a place to track the
repos it owns. This patch adds a yaml file to track User
Committee repos.

A User Committee repo was created prior to this patch and is
currently in stackforge. The repo will be renamed to live in
the openstack/ namespace and the contents of this yaml file
will be updated to reflect its location when it changes.

Change-Id: If6dbd44266b95a189cd1746bfb245cf1c31616d0
Reference: https://review.openstack.org/#/c/185106/