Updated openstack/openstack Project: openstack/ceilometer  53e70ea46db29a291e20a8c6c4fac4edda939d9a
Merge tag '2014.2'

This is a null-merge of the 2014.2 release tag back into the
master branch. It contains no actual changes to the master branch,
regardless of how our code review system's UI represents it. Please
ask in #openstack-infra if you have any questions, and otherwise try
to merge this as quickly as possible to avoid later conflicts on the
master branch.

Change-Id: Iaed05cac703604639608e5fd203cffc2ed474e34