Updated openstack/openstack Project: openstack/ceilometer  9f2eb2118d30fa1c08607d40c2310f78f49fbf45
Handle database failures on api startup

Storage layer connections are initialized
for metering, alarm and event in Pecan hook
middleware.
If any of the backends are not responding,
the connections are retried till maximum
retries and the exception propagated to the
Ceilometer-api which fails to startup.

This fix logs and handles the exception in
the Pecan hook and allows all connection
types to be initialized.

Closes-Bug: 1456944

Change-Id: Id70693dafc823927b3d72e0371e3e55d4e5588ce