Updated openstack/openstack Project: openstack/nova  97a97205a980459bae1f61aec3d4c7e0bec1e9c2
libvirt: fix live migration handling of disk_info

Previous commmit:

  commit a7a0b4e542a65b3ba3650db2b7560eb587b9606c
  Author: Joel Coffman <joel.coffman@jhuapl.edu>
  Date:   Fri Apr 24 16:14:25 2015 -0400

    libvirt: Remove unnecessary JSON conversions

Changes the '_create_images_and_backing' method to take a
dict instead of a JSON string for the 'disk_info' parameter.
Unfortunately it only updated one of the callers, so the
migration code was still passing a JSON string.

This wasn't caught because all the unit tests were broken,
mistakenly passing a dict instead of a json string to the
pre_live_migration method, and the _create_images_and_backing
method was stubbed out with a mock, so never ran. Add a
check on the parameters to the _create_images_and_backing
mock to ensure it is given a dict, and not a json string.

Change-Id: I6a8f009bb21bce9e6408ebc4508094bdcbef05c5
Closes-bug: #1463747