Updated openstack/openstack Project: openstack/glance  b3a824bb4b07426e02b8986ed4a9c0e10c8d5fe7
Fixed error message for negative values of min_disk and min_ram

In case of v1 api if min_disk or min_ram is passed with a negative value
to the image create command, it returns 400 response with error message
formatted to HTML entity.

Eliminated special characters (>=) from error message, so it will not
cause any encoding issues. Also with this change the error message for
v1 api will be identical to v2 api.

Closes-Bug: 1461776
Change-Id: I71a3cc3ed6bf14f3c0273df053667f1cbf99e98f