Updated openstack/openstack Project: openstack-infra/project-config  0addabc8d62ae92e6304fe192ea538765db209d8
Disable future parser check on puppet-swift (juno)

The puppet-future-syntax check runs a puppet 3 parser with the future
parser option turned on in preparation for migrating to puppet 4.
Unfortunately, it fails on the stable/juno branch of the puppet-swift
module[1]. The fix has been committed in master[2] but cannot be
backported because it would break the API. The only way to continue
development on the stable/juno branch is to disable this check and
accept that the Juno version of the puppet-swift module is not
future-parser or puppet 4 compatible.

[1] http://logs.openstack.org/52/188952/1/check/gate-puppet-swift-puppet-syntax-future/1207ef3/console.html.gz#_2015-06-05_22_21_20_293
[2] https://review.openstack.org/#/c/137187/

Change-Id: Id4295f0410b5f9f7acc568694811cd3dba6cae52