Updated openstack/openstack Project: openstack/tripleo-puppet-elements  5c945d3febeda8d3088cf6e2bcce1fdc521e1879
Preserve data types in Hiera where possible

We didn't preserve the data types from Heat metadata when writing hiera
files. We treated the metadata as raw YAML, unless it was multiline, in
which case we wrapped it in quotes.

This commit fixes those things as much as possible:

* Proper JSON escaping is preserved for multi-line strings.

* Other data types such as arrays and hashes are properly passed into
  hiera, regardless if their string representation would be single-line
  or multi-line.

However, single-line strings are still treated as raw YAML, because
tripleo-heat-templates already depend on that behavior (instead of
having arrays in Heat metadata, we have strings there formatted as
arrays).

Change-Id: Ie984f3c7782687235cdc2d72ef1f94af89dc3ed4
Closes-Bug: #1462369