Updated openstack/openstack Project: openstack/tuskar  0bfaf4c7afbd66833f84a73384a74d40beb97a8e
Propagate default values for parameters from the master seed

Some parameters like ServiceNetMap are defaulted in the master_seed
template. If there is no default set for this parameter in any of
the role templates where it is used, propagate the default value
from the master_seed to that role (overcloud-without-mergepy.yaml).
The default value is also included in the environment file

Change-Id: I1ff97ccabf6be6afba03fc08f817c0d98f554ed2