Updated openstack/openstack Project: openstack/openstack-manuals  86ea72501a14eb5f97067ab0c624f5b202ec14bc
Instructions to resolve an openSUSE 13.2 dependency conflict

Adds a note with instructions to respond to a pattern dependency
conflict when a minimal server install is used.

Modified: replaced the note with a simpler instruction to remove
the offending package in the Basic Environment section of the
Install Guide.

Change-Id: I2ad6c8cf7ce76fa42950e51f267224cd8142199c
Closes-bug: #1462624
Backport: Kilo