Updated openstack/openstack Project: openstack/ceilometer  d488d3683570a20ab4fc9aafe07f6490fcbbd610
Merge tag '2015.1.0'

This is a null-merge of the 2015.1.0 release tag back into the master
branch so that the 2015.1.0 tag will appear in the git commit history of
the master branch. It contains no actual changes to the master branch,
regardless of how our code review system's UI represents it. Please
ask in #openstack-infra if you have any questions, and otherwise try
to merge this as quickly as possible to avoid later conflicts on the
master branch.

Change-Id: I3c81175c88744eed642b2448ae92ad822f0b9b36