Updated openstack/openstack Project: openstack-infra/elastic-recheck  0730037103b31a26b821605c690a6e97dd8ccf2d
Expand fingerprint for 1463478

Another manifestation of bug 1463478 is ceilometer-api not starting up.
But we don't collect ceilometer logs in logstash, so fingerprint on the
grenade.sh output.

Expanding this fingerprint to remove hits from the unclassified page to
better to get a better sense of what are classification rate actually is.

Change-Id: Ib88a225b3be3f0868f4566005dee7fc08f5a8828
Related-Bug: #1463478