Updated openstack/openstack Project: openstack/nova  8204100fd50c4b86607499a3c15b0d0f82a013d0
libvirt: make default_device_names DRY-er

Since we now have a reusable method to call to get the block_device_info
structure that most driver code is designed to work with, we don't need
to re-write it in the blockinfo code - we can just use it.

In addition to this - this fixes a (yet unreported) issue that libvirt
would not consider blank volumes in it's device assignments.

Related-bug: 1231874
Partial-bug: 1452224

Change-Id: I4b9a6fd1b08ff787fdd1226f533f4181fe44b7e9