Updated openstack/openstack Project: openstack/nova  a674d4f3f3682103433ec7164e50b8352ec2eb80
Link to microversion history in docs

Make the API microversion history more accessible by adding it to nova's
doc landing page.

Making sure it is easy to discover what the difference between each
microversion is fundamental piece of the microversion puzzle, so lets
not require end users to have to look through git.

rename api_microversions.rst to api_microversion_dev.rst to clarify that
it is about development and versus history.

Change-Id: I270944bf9739b113a43b932948fdbf83e449603b