Updated openstack/openstack Project: openstack/nova  13703a2ece20228d7eae916925563942a3d33deb
Add the swap and ephemeral BDMs if needed

After preparing the block device mapping for creating new instance(s),
check if any ephemeral and/or a swap disks have been already defined. If
that is not the case and the instance type requires them, create the
needed ephemeral and/or swap BDMs.

Closes-Bug: #1297325
Related-Bug: #1457527
Change-Id: I44b30625cf1023d20ebec5e38f46b7b8dab325f6