Updated openstack/openstack Project: openstack/nova  b873badc0b60a3bf53cac03a6e220d151953911b
libvirt: Don't try to confine a non-NUMA instance

Even though it seemed like a good idea at the time, it turns out that
choosing a NUMA node at random can lead to placements that are
suboptimal.

Since the benefit of the current approach is unclear, and the downside
is actually hurting users - it is best to remove it (we would be no
worse off than before Nova was even NUMA aware), and think about
optimizitaions going forward, if any.

Change-Id: Idd4d5c236067fc40ed3a5049f3c1a58b75f9b5e5
Closes-bug: #1461777