Updated openstack/openstack Project: openstack/heat  4d42df500a3ef7a03dccd230e26727335d4e13b1
Move mistral resources in-tree

This change relocates the mistral resources from the contrib area into the main
resource tree. It was originally added to contrib/ because of the project's
incubation status, and more specifically because the client is not in the
global-requirements.txt file. However, when this was discussed at the summit in
Vancouver, the decision was to move the resources to main tree but skip
registration if the client is not installed. This will save users from the
trouble of installing it as a plugin.

Change-Id: I6eeef5fa2b080df610e52620d2b935450d8d49e3