Updated openstack/openstack Project: openstack-infra/shade  604c27fb54708e5d6abc1a462527b86ef4d9cbc3
Add CRUD methods for Keystone domains

We're going to call these identity domains. I'm very tempted to call
them realms, since that's the name that would be most specific and since
Domain is clearly a thing owned by the Domain Name Service, but I'll
stop using this commit message as a rant now.

Co-Authored-By: Haneef Ali <haneef.ali@hp.com>
Change-Id: I2dd10aa7081fdb1d0bf21aa266e5a707e965f055