Updated openstack/openstack Project: openstack/heat-translator  e5827c8973e1dc4aa05e27ddd4a3c5034415497d
Installer scripts should not run simultaneously on the same server

Due to Linux's limitation of apt-get and yum utilities that cannot
run multiple sessions in parallel, SoftwareDeployment resources
that install some package, are deployed to the same server, and do
not have any dependency between them could fail installing those
packages. We update the scripts to verify no other install session
is in progress before attemting an install. This needs to be done
in any future installer script that makes use of such utilities.

Change-Id: Iea19fe27acffda7c54db801b728a66ea427e8662
Closes-Bug: #1455251
Closes-Bug: #1463595