Updated openstack/openstack Project: openstack/keystonemiddleware  666150149889d1d167fb3fba06d43f4b07809503
Refactor request methods onto request object

Take advantage of webob Request objects to move methods that deal with
manipulating the request onto the request.

This removes a debug statement that prints a static list of headers that
will be removed because we no longer keep the static list. This list was
not really useful to debugging as the list doesn't change.

Change-Id: I8601cd7ff9cea778911cbfc306dc03089530bc9d