Updated openstack/openstack Project: openstack/horizon  ea2b14fae4332952829a66a251dabaca20003f2e
corrected the wrong url in admin instance detail

In admin instance detail page, the image url pointing to image detail page in project panel.

This behavior is wrong, it should be link to admin panel only.

This patch corrected the url to point to the admin panel image detail page.

Change-Id: I4ec2687cc208e61b2be00dc86e42f95fa8a1a363
Closes-Bug: #1463312