Updated openstack/openstack Project: openstack/nova-specs  a24d53f65e7146eeebee121e6eaede8ab8038281
Add liberty priorities

At the summit we agreed the priorities for the liberty release.
This records the outcome of that discussion, and allows those unable
to be at the summit to be involved in  that discussion.

In addition, I have added some information about the most important
things that were discussed, but don't yet have a firm enough plan
to go on the list of priorities.

The raw discussion notes can be found here:
https://etherpad.openstack.org/p/YVR-nova-liberty-priorities

Change-Id: I806b35a29f70a09fd06692adff6d0af9ef486fe3