Updated openstack/openstack Project: openstack-infra/puppet-zanata  9e0edabed4cfc5156b698444e30f8937b3071c02
Correctly determine the server id

Due to using an undefined variable when determining the server id,
the client manifest was writing an incorrectly formatted zanata.ini
file. Furthermore, the intent of the change did not go far enough,
since it would not strip off the URI scheme or port. Due to the
complexities of parsing URIs in regular expressions, use a Puppet
function for the heavy lifting.

Change-Id: I754ee54f805c91f5548b2cf270b23c68eed3959c