Updated openstack/openstack Project: openstack/requirements  21fa5439269fe07978a04f252ebac11cd1b8f9d6
Move update.py into the python package.

A symlink is kept to update.py to keep this working for infra until we
update the propose scripts to try the new location and fall back.

Until that is done and the symlink removed, update.py cannot refer to
other modules (though they can refer to it).

Change-Id: If878bc2bb426c0e5478ed031f1144e26f05e85a2