Updated openstack/openstack Project: openstack/zaqar  f1d153a5880fea3ae290047f45b1816b5b50546d
Merge tag '2015.1.0'

This is a null-merge of the 2015.1.0 release tag back into the master
branch so that the 2015.1.0 tag will appear in the git commit history of
the master branch. It contains no actual changes to the master branch,
regardless of how our code review system's UI represents it. Please
ask in #openstack-infra if you have any questions, and otherwise try
to merge this as quickly as possible to avoid later conflicts on the
master branch.

Change-Id: Ic5b95fdfa8949e85ca8601d39137cc4e4dcbb4cd