Updated openstack/openstack Project: openstack/tempest  4209ecfa60b96b35c8c1c74fcf4e0b34d96ae4cb
Fix L3AgentSchedulerTestJSON failure for dvr routers

In L3AgentSchedulerTestJSON test class, the resource
setup method has special handling for the dvr routers,
since DVR Routers require a valid subnet with port for
the routers to be scheduled.

But in the case of Non-DVR Routers this was not required.

The resource cleanup method was missing in this test class.
But the Super class resource cleanup method was not aware
about the DVR Router and existence of the router owned ports
and hence was trying to delete the port before it removed
the router.

This patch fixes the problem by creaing a resource cleanup
method in L3AgentSchedulerTestJSON and cleans up the
interface and subnet added for the DVR Router.

Change-Id: I0fbc02a2bc8badbb1792870c34495b913572af0a
Closes-Bug: #1462150