Updated openstack/openstack Project: openstack/neutron  7782d806fb5e8044840b81608f05b4e056dbc8ff
Document existence of br-tun and br-int in the OVS agent

Question about the use of the two bridges has come up in the past
multiple times, so let's fill the gap in the developer documentation.

A user-facing documentation patch will have to follow up, if we want
to be very thorough.

Change-Id: I6dac0f9bdaf7b3b7bff8745d4103ccc71df61a0a