Updated openstack/openstack Project: openstack/python-saharaclient  26cf2237fc27f98bd05eaf4648acfdb36d685ff1
Pass OpenStack environment variables to client tests

tox 2 does not pass the environment by default, this is
a change from tox 1.6. Instead, environment variables
may be passed by means of the "passenv" variable in
tox.ini. This mechanism supports globbing, so "OS_*"
can be used to set the necessary environment variables
for the python-saharaclient.

Change-Id: Ibe9055dd4d4bc1a243db5198a9e5e35e5f93f628