Updated openstack/openstack Project: openstack/python-zaqarclient  e9a8d01cf8995084e825cad1538b7e1cda7e78aa
Added support to pools and flavors

Python-zaqarclient had missing features for v1.1 which involved pools and flavors that
needed to be added.

I added support to pools by adding a 'pool_update' and a 'pool_list' function,
and add support to flavors by writing a 'flavor_update' and a 'flavor_list'
function.

Change-Id: I640d17c64be6448283c53ae0dbea6543484ff7a4
Depends-on: I36445838f532e9f282a4f839a0f583257c1f13d9
Partially-implements: # blueprint api-v1.1