Updated openstack/openstack Project: openstack/neutron-specs  b37a5be0426dcd3e971595af9797a4ed34fc5bcb
Neutron API evolution strategy

This specification discusses, in a very boring and pedant
way, a plan for moving away from extensions as a way to
evolve the API and introduce micro versioning as it has
already been done for Nova and Ironic.

Since, it is very boring, read only if properly caffeinated
(or if you are trying to get asleep).

DocImpact
ApiImpact

Change-Id: I7674147172baec42cf3ee03f13ef4e4255586447