Updated openstack/openstack Project: openstack/horizon  023bd3a90f47501e32d1c9c420260406ccecb295
ngReorg - move core HTML from framework to dashboard

This patch moves the base.html file and related files (where appropriate)
to the dashboard. In short:

- base.html its *dashboard configuration* components _header.html,
  _stylesheets.html, horizon/_conf.html, horizon/_custom_head_js.html,
  horizon/_custom_meta.html and horizon/_scripts.html all move to the
  dashboard
- other components in base.html which are effectively widgets or
  default content (horizon/_messages.html, horizon/common/_sidebar.html,
  horizon/common/_page_header.html,
  horizon/client_side/_script_loader.html) stay in the framework

The framework test suite also includes a need for a base.html which can
construct *part* of the application's base.html structure, and this is
created in horizon/test/templates/base.html (this file is only used to
test Django templating, so does not need most of the content in the real
base.html).

Change-Id: I9b3c4f21036d08e1d79aeab1ab35d47e3eedd3a3
Partial-Bug: #1458697