Updated openstack/openstack Project: openstack/nova  d81f6eb4d32f925459ce3fa335050c3bcfc06c65
Remove cell policy check

we already got policy checks for create,update,delete
and sync_instances of cell, there is no need to check the
general policy of cell and it might lead to inconsistency
problem.

The v2 code is not touched because v2.1 is current using.

UpgradeImpact

Partially implements bp nova-api-policy-final-part

Change-Id: I05ea8b0f3b093b7dbfbe037c21e224acc3c605ad