Updated openstack/openstack Project: openstack-infra/devstack-gate  ab3712d288c8172fd489068c6399566f272509b0
Add support for DEVSTACK_GATE_SETTINGS

This patch adds support for DEVSTACK_GATE_SETTINGS.  A job definition
can set this variable with a path to a file to source after the
environment has been set up.  This is useful for projects to be able
to set custom environment variables from a file in their project
instead of maintaining them in project-config.

Note that this is not possible in the pre_test_hook as that runs in a
sub-shell, so it's not possible to set environment variables that
remain.

Change-Id: I80742d498f4f21006eed16a82d568b496957797f
Signed-off-by: Russell Bryant <rbryant@redhat.com>