Updated openstack/openstack Project: openstack/zaqar  b5379b78838787062087e2229b6e29382b14e271
Try to use a default pool when creating a queue

We recently enabled default pools creation, which will use the default
`message_storage` if present. This patch extends that to queue's
creation where we try to register a queue in the catalogue but we fail
if no pools were found.

With this patch, the failure will happen after checking if there's a
default pool that can be used.

In addition to the above, this commit adds a new option
`enable_virtual_pool` that makes default pools registration optional.
This is to avoid breaking backwards compatibility on the deployments
already running Zaqar.

DocImpact

Closes-bug: #1462008
Closes-bug: #1463110
Change-Id: Ieb8d2671066fc909f422bd74fd20070bc0f9cf11