Updated openstack/openstack Project: openstack-dev/devstack  7a6f0b6bffd412b7d647af66dea3d972370e9440
guru meditation report for nova-compute in worlddump

Nova-compute is hanging in the multinode test, and its difficult to
figure out why. So trigger a guru meditation report for nova-compute in
worlddump so we can see what nova-compute is doing when it is hung.

Having a hung nova-compute causes tempest to fail and
I035fe8e3333034e44b403ed0f986220ab5b0e57a runs worlddump whenever
tempest fails.

Bug 1462305 is one of the last issues left before the multinode job is
stable enough to gate on, and this patch should make it much easier to
debug.

Change-Id: I87d7536b5992c47b8082684cc662f953113fd1a8
Related-Bug: #1462305