Updated openstack/openstack Project: openstack/python-cinderclient  c54a3214106a137a1edfe4c3c2a975335a7b481f
Fix condition in CheckSizeArgForCreate parser action

CheckSizeArgForCreate checks that size is specified when snapshot
or source volume aren't. But when size is 0, CheckSizeArgForCreate
action works wrong and generates error: 'Size is a required
parameter if snapshot or source volume is not specified', meanwhile
user expected to see: 'Volume size '0' must be greater than 0'.

Change-Id: I164970a600d6e86bd7076dd485f676a703f5e487
Closes-Bug: #1454276