Updated openstack/openstack Project: openstack/oslo.vmware  8b0cd2604b34c21f85f14c42a0cf619aaa151a76
Update progress lease every 60 seconds

Currently the NFC lease is updated after every written/read chunk. This
is very inefficient because we make many API calls when transfering
large images. The progress can be updated every minute or so which
significantly decreases the overall transfer time.

For example, when importing 900MB streamOptimized image, the overall
time is decreased by 50% with this patch. The improvement for larger
images or slower connections will be even better.

Change-Id: Iba7754e88a43cebb4378a08d2523cfe9645f18e7