Updated openstack/openstack Project: openstack-infra/git-review  d1266ee0a5a14d8c7788337785c097a1201fa5ec
Add reviewers on upload

Add the --reviewers argument which allows specifying reviewers to be
added to each patch uploaded.  This simplifies the upload process by not
requiring an additional step of adding reviewers to changes.

Gerrit supports specifying reviewers to be added to changes by appending
a list of r='email' options to the refspec being pushed to.  For
example, from the Gerrit 'Uploading Changes' documentation:

git push tr:kernel/common HEAD:refs/for/experimental%r=a@a.com,cc=b@o.com

The --reviewers argument can be passed multiple reviewers to add to the
patch set.  Reviewers containing whitespace are rejected, as whitespace
cannot be added to the refspec.

Change-Id: I8c2f9453a90dd78aa47f7996f2502f9f6cf2d66d