Updated openstack/openstack Project: openstack/taskflow  47572341a71bc7eda78fa85a4b547eee11889838
Adding a revert_all option to retry controllers

When a retry controller is added to a nested flow, with the exception of
AlwaysRetryAll, existing retry controllers do not cause a revert of
predecessor tasks in the surrounding flow.  This is due to the fact that
these retry controllers always return RETRY.

The revert_all option, when set True, will result in the retry
controller returning REVERT_ALL rather than REVERT, which will cause
predecessor tasks in the surrounding flow to also be REVERTed.

Change-Id: I2d867b2d05e8559121ec48fd7249f15078450532