Updated openstack/openstack Project: openstack-dev/devstack  97c77b90e2064647b9968dd74fb378c584ece61f
Set Neutron api_workers to $API_WORKERS

Change [1] brings back this feature in Neutron, so we want to have
the ability to set the number of API workers the same way other
projects do.

However, this cause some instability, so we need to be careful on
how we bring it back.

[1] https://review.openstack.org/#/c/191127/

Closes-Bug: #1432189
Related-bug: #1432065

Change-Id: Id4986a49d33fa4b8a7291150488665e200525dac
Co-authored-by: Russell Bryant <rbryant@redhat.com>