Updated openstack/openstack Project: openstack/nova  e9b139096a7faf5ff344ecbbf916f7dd12a3fc56
Cleanup wording for the disable_libvirt_livesnapshot workaround option

The config options go in the docs so the help on them should be clear,
especially for workaround options since they are for known issues and
people have to opt-in to using them.

This clears up the wording in the disable_libvirt_livesnapshot option.

Note the change from 'mechanism to disable' to 'mechanism to enable'
since the feature is disabled by default, so the option is actually
providing the mechanism to enable the feature if you want to use it
provided the known issues with libvirt 1.2.2.

Change-Id: Ibb42e0b1ca7c6991ed1b189455b5279f9b12cfff