Updated openstack/openstack Project: openstack/tempest  b93e67690c35285e1fc01bfa02f7c0f9fe079765
Remove key 'absolute' from limits client

limits client should be just for getting a resource "limits",
it is not necessary to specify key 'absolute' at all as the library
method.
In addition, this patch removes get_specific_absolute_limit because
the method was just wrapper. That should be implemented on test side.

Change-Id: I5c0ffc136939d0aeb8574c180353715f2a0e2665